Split controller
