+++ src/routes/owners/owners.controller.v1.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, Param } from '@nestjs/common';
-import { ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
+import { ApiOkResponse, ApiTags } from '@nestjs/swagger';
 import { SafeList } from '@/routes/owners/entities/safe-list.entity';
 import { OwnersService } from '@/routes/owners/owners.service';
 import { ValidationPipe } from '@/validation/pipes/validation.pipe';
@@ -8,12 +8,13 @@ import { AddressSchema } from '@/validation/entities/schemas/address.schema';
 @ApiTags('owners')
 @Controller({
   path: '',
+  version: '1',
 })
-export class OwnersController {
+export class OwnersControllerV1 {
   constructor(private readonly ownersService: OwnersService) {}
 
   @ApiOkResponse({ type: SafeList })
-  @Get('/v1/chains/:chainId/owners/:ownerAddress/safes')
+  @Get('chains/:chainId/owners/:ownerAddress/safes')
   async getSafesByOwner(
     @Param('chainId') chainId: string,
     @Param('ownerAddress', new ValidationPipe(AddressSchema))
@@ -22,39 +23,12 @@ export class OwnersController {
     return this.ownersService.getSafesByOwner({ chainId, ownerAddress });
   }
 
-  @ApiOkResponse({
-    schema: {
-      type: 'object',
-      additionalProperties: {
-        type: 'array',
-        items: { type: 'string' },
-      },
-    },
-  })
-  @ApiOperation({ deprecated: true })
-  @Get('/v1/owners/:ownerAddress/safes')
-  async deprecated__getAllSafesByOwner(
+  @ApiOkResponse({ type: SafeList })
+  @Get('owners/:ownerAddress/safes')
+  async getAllSafesByOwner(
     @Param('ownerAddress', new ValidationPipe(AddressSchema))
     ownerAddress: `0x${string}`,
   ): Promise<{ [chainId: string]: Array<string> }> {
     return this.ownersService.deprecated__getAllSafesByOwner({ ownerAddress });
-  }
-
-  @ApiOkResponse({
-    schema: {
-      type: 'object',
-      additionalProperties: {
-        type: 'array',
-        items: { type: 'string' },
-        nullable: true,
-      },
-    },
-  })
-  @Get('/v2/owners/:ownerAddress/safes')
-  async getAllSafesByOwner(
-    @Param('ownerAddress', new ValidationPipe(AddressSchema))
-    ownerAddress: `0x${string}`,
-  ): Promise<{ [chainId: string]: Array<string> | null }> {
-    return this.ownersService.getAllSafesByOwner({ ownerAddress });
   }
 }
