+++ src/routes/owners/owners.controller.v2.spec.ts
@@ -33,8 +33,10 @@ import { TestPostgresDatabaseModuleV2 } from '@/datasources/db/v2/test.postgres-
 import { TestTargetedMessagingDatasourceModule } from '@/datasources/targeted-messaging/__tests__/test.targeted-messaging.datasource.module';
 import { TargetedMessagingDatasourceModule } from '@/datasources/targeted-messaging/targeted-messaging.datasource.module';
 import { rawify } from '@/validation/entities/raw.entity';
-import { LoggingService } from '@/logging/logging.interface';
-import type { ILoggingService } from '@/logging/logging.interface';
+import {
+  type ILoggingService,
+  LoggingService,
+} from '@/logging/logging.interface';
 
 describe('Owners Controller (Unit)', () => {
   let app: INestApplication<Server>;
@@ -77,385 +79,6 @@ describe('Owners Controller (Unit)', () => {
 
   afterAll(async () => {
     await app.close();
-  });
-
-  describe('GET safes by owner address', () => {
-    it(`Success`, async () => {
-      const chainId = faker.string.numeric();
-      const ownerAddress = faker.finance.ethereumAddress();
-      const chainResponse = chainBuilder().with('chainId', chainId).build();
-      const transactionApiSafeListResponse = {
-        safes: [
-          faker.finance.ethereumAddress(),
-          faker.finance.ethereumAddress(),
-          faker.finance.ethereumAddress(),
-        ],
-      };
-      networkService.get.mockResolvedValueOnce({
-        data: rawify(chainResponse),
-        status: 200,
-      });
-      networkService.get.mockResolvedValueOnce({
-        data: rawify(transactionApiSafeListResponse),
-        status: 200,
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/chains/${chainId}/owners/${ownerAddress}/safes`)
-        .expect(200)
-        .expect({
-          // Validation schema checksums addresses
-          safes: transactionApiSafeListResponse.safes.map((safe) =>
-            getAddress(safe),
-          ),
-        });
-    });
-
-    it('Failure: Config API fails', async () => {
-      const chainId = faker.string.numeric();
-      const ownerAddress = faker.finance.ethereumAddress();
-      const error = new NetworkResponseError(
-        new URL(
-          `${safeConfigUrl}/v1/chains/${chainId}/owners/${ownerAddress}/safes`,
-        ),
-        {
-          status: 500,
-        } as Response,
-      );
-      networkService.get.mockRejectedValueOnce(error);
-
-      await request(app.getHttpServer())
-        .get(`/v1/chains/${chainId}/owners/${ownerAddress}/safes`)
-        .expect(500)
-        .expect({
-          message: 'An error occurred',
-          code: 500,
-        });
-
-      expect(networkService.get).toHaveBeenCalledTimes(1);
-      expect(networkService.get).toHaveBeenCalledWith({
-        url: `${safeConfigUrl}/api/v1/chains/${chainId}`,
-      });
-    });
-
-    it('Failure: Transaction API fails', async () => {
-      const chainId = faker.string.numeric();
-      const ownerAddress = faker.finance.ethereumAddress();
-      const chainResponse = chainBuilder().with('chainId', chainId).build();
-      networkService.get.mockResolvedValueOnce({
-        data: rawify(chainResponse),
-        status: 200,
-      });
-      const error = new NetworkResponseError(
-        new URL(
-          // ValidationPipe checksums ownerAddress param
-          `${chainResponse.transactionService}/v1/chains/${chainId}/owners/${getAddress(ownerAddress)}/safes`,
-        ),
-        {
-          status: 500,
-        } as Response,
-      );
-      networkService.get.mockRejectedValueOnce(error);
-
-      await request(app.getHttpServer())
-        .get(`/v1/chains/${chainId}/owners/${ownerAddress}/safes`)
-        .expect(500)
-        .expect({
-          message: 'An error occurred',
-          code: 500,
-        });
-
-      expect(networkService.get).toHaveBeenCalledTimes(2);
-      expect(networkService.get).toHaveBeenCalledWith({
-        url: `${safeConfigUrl}/api/v1/chains/${chainId}`,
-      });
-      expect(networkService.get).toHaveBeenCalledWith({
-        // ValidationPipe checksums ownerAddress param
-        url: `${chainResponse.transactionService}/api/v1/owners/${getAddress(ownerAddress)}/safes/`,
-      });
-    });
-
-    it('Failure: data validation fails', async () => {
-      const chainId = faker.string.numeric();
-      const ownerAddress = faker.finance.ethereumAddress();
-      const chainResponse = chainBuilder().with('chainId', chainId).build();
-      const transactionApiSafeListResponse = {
-        safes: [
-          faker.finance.ethereumAddress(),
-          faker.number.int(),
-          faker.finance.ethereumAddress(),
-        ],
-      };
-      networkService.get.mockResolvedValueOnce({
-        data: rawify(chainResponse),
-        status: 200,
-      });
-      networkService.get.mockResolvedValueOnce({
-        data: rawify(transactionApiSafeListResponse),
-        status: 200,
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/chains/${chainId}/owners/${ownerAddress}/safes`)
-        .expect(502)
-        .expect({ statusCode: 502, message: 'Bad gateway' });
-    });
-  });
-
-  describe('deprecated: GET all safes by owner address', () => {
-    it('Success for singular chain page', async () => {
-      const ownerAddress = faker.finance.ethereumAddress();
-
-      const chainId1 = faker.string.numeric();
-      const chainId2 = faker.string.numeric({ exclude: [chainId1] });
-
-      const chain1 = chainBuilder().with('chainId', chainId1).build();
-      const chain2 = chainBuilder().with('chainId', chainId2).build();
-
-      const safesOnChain1 = [
-        faker.finance.ethereumAddress(),
-        faker.finance.ethereumAddress(),
-        faker.finance.ethereumAddress(),
-      ];
-      const safesOnChain2 = [
-        faker.finance.ethereumAddress(),
-        faker.finance.ethereumAddress(),
-        faker.finance.ethereumAddress(),
-      ];
-
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${safeConfigUrl}/api/v1/chains`: {
-            return Promise.resolve({
-              data: rawify(
-                pageBuilder()
-                  .with('results', [chain1, chain2])
-                  .with('next', null)
-                  .build(),
-              ),
-              status: 200,
-            });
-          }
-
-          case `${safeConfigUrl}/api/v1/chains/${chainId1}`: {
-            return Promise.resolve({
-              data: rawify(chain1),
-              status: 200,
-            });
-          }
-
-          case `${safeConfigUrl}/api/v1/chains/${chainId2}`: {
-            return Promise.resolve({
-              data: rawify(chain2),
-              status: 200,
-            });
-          }
-
-          // ValidationPipe checksums ownerAddress param
-          case `${chain1.transactionService}/api/v1/owners/${getAddress(ownerAddress)}/safes/`: {
-            return Promise.resolve({
-              data: rawify({ safes: safesOnChain1 }),
-              status: 200,
-            });
-          }
-
-          case `${chain2.transactionService}/api/v1/owners/${getAddress(ownerAddress)}/safes/`: {
-            return Promise.resolve({
-              data: rawify({ safes: safesOnChain2 }),
-              status: 200,
-            });
-          }
-
-          default: {
-            return Promise.reject(`No matching rule for url: ${url}`);
-          }
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/owners/${ownerAddress}/safes`)
-        .expect(200)
-        .expect({
-          // Validation schema checksums addresses
-          [chainId1]: safesOnChain1.map((safe) => getAddress(safe)),
-          [chainId2]: safesOnChain2.map((safe) => getAddress(safe)),
-        });
-    });
-
-    it('Success for multiple chain pages', async () => {
-      const ownerAddress = faker.finance.ethereumAddress();
-
-      const chainId1 = faker.string.numeric();
-      const chainId2 = faker.string.numeric({ exclude: [chainId1] });
-
-      const chain1 = chainBuilder().with('chainId', chainId1).build();
-      const chain2 = chainBuilder().with('chainId', chainId2).build();
-
-      const chainsUrl = `${safeConfigUrl}/api/v1/chains`;
-      const offset = 1;
-      const chainsPage1 = pageBuilder()
-        .with('results', [chain1])
-        .with('next', limitAndOffsetUrlFactory(undefined, offset, chainsUrl))
-        .build();
-      const chainsPage2 = pageBuilder()
-        .with('results', [chain2])
-        .with('next', null)
-        .build();
-
-      const safesOnChain1 = [
-        faker.finance.ethereumAddress(),
-        faker.finance.ethereumAddress(),
-        faker.finance.ethereumAddress(),
-      ];
-      const safesOnChain2 = [
-        faker.finance.ethereumAddress(),
-        faker.finance.ethereumAddress(),
-        faker.finance.ethereumAddress(),
-      ];
-
-      networkService.get.mockImplementation(({ url, networkRequest }) => {
-        if (url === chainsUrl && !networkRequest!.params!.offset) {
-          return Promise.resolve({
-            data: rawify(chainsPage1),
-            status: 200,
-          });
-        }
-        if (url === chainsUrl && networkRequest!.params!.offset === offset) {
-          return Promise.resolve({
-            data: rawify(chainsPage2),
-            status: 200,
-          });
-        }
-        if (url === `${safeConfigUrl}/api/v1/chains/${chainId1}`) {
-          return Promise.resolve({
-            data: rawify(chain1),
-            status: 200,
-          });
-        }
-
-        if (url === `${safeConfigUrl}/api/v1/chains/${chainId2}`) {
-          return Promise.resolve({
-            data: rawify(chain2),
-            status: 200,
-          });
-        }
-
-        // ValidationPipe checksums ownerAddress param
-        if (
-          url ===
-          `${chain1.transactionService}/api/v1/owners/${getAddress(ownerAddress)}/safes/`
-        ) {
-          return Promise.resolve({
-            data: rawify({ safes: safesOnChain1 }),
-            status: 200,
-          });
-        }
-
-        if (
-          url ===
-          `${chain2.transactionService}/api/v1/owners/${getAddress(ownerAddress)}/safes/`
-        ) {
-          return Promise.resolve({
-            data: rawify({ safes: safesOnChain2 }),
-            status: 200,
-          });
-        }
-
-        return Promise.reject(`No matching rule for url: ${url}`);
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/owners/${ownerAddress}/safes`)
-        .expect(200)
-        .expect({
-          // Validation schema checksums addresses
-          [chainId1]: safesOnChain1.map((safe) => getAddress(safe)),
-          [chainId2]: safesOnChain2.map((safe) => getAddress(safe)),
-        });
-    });
-
-    it('Failure: Config API fails', async () => {
-      const ownerAddress = faker.finance.ethereumAddress();
-
-      networkService.get.mockImplementation(({ url }) => {
-        if (url === `${safeConfigUrl}/api/v1/chains`) {
-          const error = new NetworkResponseError(
-            new URL(`${safeConfigUrl}/api/v1/chains`),
-            {
-              status: 500,
-            } as Response,
-          );
-          return Promise.reject(error);
-        }
-        return Promise.reject(`No matching rule for url: ${url}`);
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/owners/${ownerAddress}/safes`)
-        .expect(500)
-        .expect({
-          message: 'An error occurred',
-          code: 500,
-        });
-
-      expect(networkService.get).toHaveBeenCalledTimes(1);
-      expect(networkService.get).toHaveBeenCalledWith({
-        url: `${safeConfigUrl}/api/v1/chains`,
-        networkRequest: {
-          params: { limit: ChainsRepository.MAX_LIMIT, offset: 0 },
-        },
-      });
-    });
-
-    it('Failure: data validation fails', async () => {
-      const ownerAddress = faker.finance.ethereumAddress();
-
-      const chainId = faker.string.numeric();
-
-      const chain = chainBuilder().with('chainId', chainId).build();
-
-      const safesOnChain = [
-        faker.finance.ethereumAddress(),
-        faker.number.int(),
-        faker.finance.ethereumAddress(),
-      ];
-
-      networkService.get.mockImplementation(({ url }) => {
-        switch (url) {
-          case `${safeConfigUrl}/api/v1/chains`: {
-            return Promise.resolve({
-              data: rawify({
-                results: [chain],
-              }),
-              status: 200,
-            });
-          }
-
-          case `${safeConfigUrl}/api/v1/chains/${chainId}`: {
-            return Promise.resolve({
-              data: rawify(chain),
-              status: 200,
-            });
-          }
-
-          case `${chain.transactionService}/api/v1/owners/${ownerAddress}/safes/`: {
-            return Promise.resolve({
-              data: rawify({ safes: safesOnChain }),
-              status: 200,
-            });
-          }
-
-          default: {
-            return Promise.reject(`No matching rule for url: ${url}`);
-          }
-        }
-      });
-
-      await request(app.getHttpServer())
-        .get(`/v1/owners/${ownerAddress}/safes`)
-        .expect(502)
-        .expect({ statusCode: 502, message: 'Bad gateway' });
-    });
   });
 
   describe('GET all safes by owner address', () => {
